Arena allocator
This allocator allocates a 4MiB arena into which all allocations are
made, and then increasingly larger arenas as earlier ones are used up.
Freeing memory in the arena is a no-op: clean all memory with
cmark_arena_reset().

In order to support realloc, we store the size of each allocation in a
size_t before the returned pointer.

The speedup is over 25% on large (benchmark-sized) inputs -- we pay a
small increase in maximum RSS (~10%) for this.