Prevent out-of-bound memory access.
Before the change, the scan_delimiters function could return max_delims+1.
The scan_delimiters function is called by extensions/strikethrough with:

18:    inline_parser, sizeof(buffer)-1, '~'

where buffer is:

12:    char buffer[101];

So scan_delimiters can return delims = 101. A few lines later, it is
used to 0-terminate the buffer with:

23:    buffer[delims] = 0;

so it effectively does buffer[101] = 0, which is an out-of-bound access.

A test was added. That's the test that I used inside Google to reproduce
the bug, thanks to internal infrastructure (sanitizers) to detect bad
memory access. Unfortunately, I couldnt' reproduce the bug with
cmark-gfm build suite. I tried `make asan test` but it gave nothing. I'm
still including the test for reference.