Fix bug where tasklist extension was using union in two ways.
The tasklist extension was using the "as" union both as a list
(in `tasklist.c:parse_node_item_prefix`) and as an opaque
(in `tasklist.c:open_tasklist_item`). This meant that strange bugs
could occur because the underlying union memory was being overwritten.

It manifested when using nested task lists indented by 4 spaces.

To fix this, I added the "checked" field to the `cmark_list` structure.
This allows the tasklist extension to use the `as.list` union member
in all its operations. This is appropriate because a tasklist item
is a list item in all essentials -- and even shares the CMARK_NODE_ITEM
type.